# api/swagger.yaml
openapi: 3.0.3
info:
  title: DevNest Portal API (OpenAPI 3.0)
  description: The DevNest Portal API endpoints definitions based on the OpenAPI 3.0
    specification.
  termsOfService: https://devnest.ro/terms-and-conditions
  contact:
    email: contact@devnest.ro
  version: 1.0.0
...
servers:
- url: https://me.devnest.ro/api/v1
...
paths:
...
  /employees:
    get:
      tags:
      - Employees
      summary: Get all employees
      description: Get all employees
      operationId: getAllEmployees
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:employees
    post:
      tags:
      - Employees
      summary: Create a new employee in the database
      description: Create a new employee in the database
      requestBody:
        description: Employee data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewEmployee'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:employees
  /employees/{employeeId}:
    get:
      tags:
      - Employees
      summary: Get employee by Id
      description: Get employee by Id
      operationId: getEmployeeById
      parameters:
      - name: employeeId
        in: path
        description: ID of the employee
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:employees
    put:
      tags:
      - Employees
      summary: Update an employee
      description: Update an employee
      parameters:
      - name: employeeId
        in: path
        description: ID of the employee
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      requestBody:
        description: Employee data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:employees
    delete:
      tags:
      - Employees
      summary: Deletes an existing employee
      description: Delete an existing employee by id
      operationId: deleteEmployeeById
      parameters:
      - name: employeeId
        in: path
        description: ID of the employee
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "204":
          description: The resource was deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:employees
    patch:
      tags:
      - Employees
      summary: Update the 'status' parameter based on whether the employee is archived
        or not.
      description: Update the 'status' parameter based on whether the employee is
        archived or not.
      parameters:
      - name: employeeId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/employees_employeeId_body'
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/employees_employeeId_body'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:employees
...
  /employees/{employeeId}/history:
    get:
      tags:
      - Employees
      summary: Get history for an employee
      description: Get history for an employee
      operationId: getHistoryForEmployee
      parameters:
      - name: employeeId
        in: path
        description: ID of the employee
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/History'
                x-content-type: application/json
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:history
...
  /feedback:
    get:
      tags:
      - Feedback
      summary: Get all feedback
      description: Get all feedback
      operationId: getAllFeedback
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeFeedback'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:feedback
  /feedback/{employeeId}:
    get:
      tags:
      - Feedback
      summary: Get overall feedback for an employee
      description: Get overall feedback for an employee
      operationId: getEmployeeOverallFeedback
      parameters:
      - name: employeeId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedEmployeeFeedback'
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:feedback
  /feedback/{employeeId}/month:
    get:
      tags:
      - Feedback
      summary: Get all monthly feedback for an employee
      description: Get all monthly feedback for an employee
      operationId: getEmployeeMonthlyFeedback
      parameters:
      - name: employeeId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MonthlyFeedback'
                x-content-type: application/json
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:feedback
    post:
      tags:
      - Feedback
      summary: Create feedback for a specific employee in a specific month
      description: Create feedback for a specific employee in a specific month
      parameters:
      - name: employeeId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonthlyFeedback'
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyFeedback'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:feedback
  /feedback/{employeeId}/month/{feedbackId}:
    get:
      tags:
      - Feedback
      summary: Return one feedback for an employee
      description: Return one feedback for an employee for a given month
      operationId: getEmployeeFeedback
      parameters:
      - name: employeeId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      - name: feedbackId
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: number
          format: int
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feedback'
                x-content-type: application/json
        "404":
          description: The specified resource was not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:feedback
  /archives/employees:
    get:
      tags:
      - Archives
      summary: Get all archived employees
      description: Get all archived employees
      operationId: getAllArchivedEmployees
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Employee'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:archives
  /archives/projects:
    get:
      tags:
      - Archives
      summary: Get all archived projects
      description: Get all archived projects
      operationId: getAllArchivedProjects
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CreatedProject'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:archives
  /settings:
    get:
      tags:
      - Settings
      summary: Get all settings
      description: Get all settings
      operationId: getAllSettings
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - read:settings
    put:
      tags:
      - Settings
      summary: Update settings
      description: Update settings
      operationId: updateSettings
      requestBody:
        description: Settings data to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Settings'
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Settings'
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - devnest_auth:
        - write:settings
components:
  schemas:
...
    CreatedProject:
      type: object
      properties:
        id:
          type: number
          format: int
        name:
          type: string
        clientName:
          type: string
        description:
          type: string
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        currency:
          type: string
          enum:
          - RON
          - EUR
          - USD
        status:
          type: boolean
      example:
        endingDate: 2000-01-23T04:56:07.000+00:00
        clientName: clientName
        name: name
        description: description
        currency: RON
        id: 0.80082819046101150206595775671303272247314453125
        startingDate: 2000-01-23T04:56:07.000+00:00
        status: true
    Employee:
      required:
      - CNP
      - city
      - firstName
      - id
      - jobTitle
      - lastName
      - personalEmail
      - phoneNumber
      - workEmail
      type: object
      properties:
        id:
          type: number
          format: int
        userId:
          type: number
          format: int
        firstName:
          type: string
        lastName:
          type: string
        CNP:
          type: string
        jobTitle:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        workEmail:
          type: string
          format: email
        personalEmail:
          type: string
          format: email
        employeeRateType:
          type: string
          enum:
          - HOUR
          - MONTH
        employeeRateValue:
          type: number
          format: int
        contractType:
          type: string
          default: PFA
          enum:
          - PFA
          - SRL
          - CIM
        currency:
          type: string
          default: EUR
          enum:
          - EUR
          - USD
          - RON
        vacationDays:
          type: number
          format: int
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        hasMedicalPackage:
          type: boolean
        status:
          type: boolean
      example:
        lastName: lastName
        CNP: CNP
        city: city
        workEmail: ""
        contractType: PFA
        jobTitle: jobTitle
        userId: 6.02745618307040320615897144307382404804229736328125
        hasMedicalPackage: true
        employeeRateType: HOUR
        personalEmail: ""
        firstName: firstName
        phoneNumber: phoneNumber
        endingDate: 2000-01-23T04:56:07.000+00:00
        vacationDays: 5.962133916683182377482808078639209270477294921875
        currency: EUR
        id: 0.80082819046101150206595775671303272247314453125
        employeeRateValue: 1.46581298050294517310021547018550336360931396484375
        startingDate: 2000-01-23T04:56:07.000+00:00
        status: true
    NewEmployee:
      required:
      - CNP
      - city
      - firstName
      - id
      - jobTitle
      - lastName
      - personalEmail
      - phoneNumber
      - workEmail
      type: object
      properties:
        userId:
          type: number
          format: int
        firstName:
          type: string
        lastName:
          type: string
        CNP:
          type: string
        jobTitle:
          type: string
        city:
          type: string
        phoneNumber:
          type: string
        workEmail:
          type: string
          format: email
        personalEmail:
          type: string
          format: email
        employeeRateType:
          type: string
          enum:
          - HOUR
          - MONTH
        employeeRateValue:
          type: number
          format: int
        contractType:
          type: string
          default: PFA
          enum:
          - PFA
          - SRL
          - CIM
        currency:
          type: string
          default: EUR
          enum:
          - EUR
          - USD
          - RON
        vacationDays:
          type: number
          format: int
        startingDate:
          type: string
          format: date-time
        endingDate:
          type: string
          format: date-time
        hasMedicalPackage:
          type: boolean
        status:
          type: boolean
...
    History:
      type: object
      properties:
        event:
          type: string
        date:
          type: string
          format: date-time
        details:
          type: string
      example:
        date: 2000-01-23T04:56:07.000+00:00
        details: details
        event: event
    EmployeeFeedback:
      type: object
      properties:
        id:
          type: number
          format: int
        name:
          type: string
        jobTitle:
          type: string
        overallRating:
          type: number
          format: float
      example:
        overallRating: 6.0274563
        jobTitle: jobTitle
        name: name
        id: 0.80082819046101150206595775671303272247314453125
    DetailedEmployeeFeedback:
      type: object
      properties:
        id:
          type: number
          format: int
        name:
          type: string
        overallRating:
          type: number
          format: float
        teamAtmosphere:
          type: number
          format: float
        workload:
          type: number
          format: float
        clientCommunication:
          type: number
          format: float
        myInvolvementInTheProject:
          type: number
          format: float
        technicalResults:
          type: number
          format: float
      example:
        overallRating: 6.0274563
        clientCommunication: 5.637377
        name: name
        myInvolvementInTheProject: 2.302136
        workload: 5.962134
        technicalResults: 7.0614014
        id: 0.80082819046101150206595775671303272247314453125
        teamAtmosphere: 1.4658129
    MonthlyFeedback:
      type: object
      properties:
        id:
          type: number
          format: int
        month:
          type: string
        teamAtmosphere:
          type: number
          format: int
        workload:
          type: number
          format: int
        clientCommunication:
          type: number
          format: int
        myInvolvementInTheProject:
          type: number
          format: int
        technicalResults:
          type: number
          format: int
        comments:
          type: string
      example:
        clientCommunication: 5.962133916683182377482808078639209270477294921875
        comments: comments
        month: month
        myInvolvementInTheProject: 5.63737665663332876420099637471139430999755859375
        workload: 1.46581298050294517310021547018550336360931396484375
        technicalResults: 2.3021358869347654518833223846741020679473876953125
        id: 0.80082819046101150206595775671303272247314453125
        teamAtmosphere: 6.02745618307040320615897144307382404804229736328125
    Feedback:
      type: object
      properties:
        id:
          type: number
          format: int
        month:
          type: string
        employeeName:
          type: string
        submittedDate:
          type: string
          format: date
        overallRating:
          type: number
          format: float
        teamAtmosphere:
          type: number
          format: int
        workload:
          type: number
          format: int
        clientCommunication:
          type: number
          format: int
        myInvolvementInTheProject:
          type: number
          format: int
        technicalResults:
          type: number
          format: int
        comments:
          type: string
      example:
        employeeName: employeeName
        overallRating: 6.0274563
        clientCommunication: 5.63737665663332876420099637471139430999755859375
        comments: comments
        month: month
        myInvolvementInTheProject: 2.3021358869347654518833223846741020679473876953125
        workload: 5.962133916683182377482808078639209270477294921875
        technicalResults: 7.061401241503109105224211816675961017608642578125
        id: 0.80082819046101150206595775671303272247314453125
        teamAtmosphere: 1.46581298050294517310021547018550336360931396484375
        submittedDate: 2000-01-23
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: code
        message: message
...
    Settings:
      type: object
      properties:
        id:
          type: number
          format: int
        userId:
          type: number
          format: int
        notifications:
          $ref: '#/components/schemas/Settings_notifications'
      example:
        id: 0.80082819046101150206595775671303272247314453125
        userId: 6.02745618307040320615897144307382404804229736328125
        notifications:
          upcomingEndingDates:
            toggled: true
            fiveDaysBefore: false
            twoWeeksBefore: true
            oneMonthBefore: false
          completeProjectSheets: true
          employeeSheetReminder: true
          individualSheets: false
...
    employees_employeeId_body:
      type: object
      properties:
        status:
          type: boolean
      example:
        status: true
    Settings_notifications_upcomingEndingDates:
      type: object
      properties:
        toggled:
          type: boolean
          default: true
        fiveDaysBefore:
          type: boolean
          default: false
        twoWeeksBefore:
          type: boolean
          default: true
        oneMonthBefore:
          type: boolean
          default: false
      example:
        toggled: true
        fiveDaysBefore: false
        twoWeeksBefore: true
        oneMonthBefore: false
    Settings_notifications:
      type: object
      properties:
        employeeSheetReminder:
          type: boolean
          default: true
        completeProjectSheets:
          type: boolean
          default: true
        individualSheets:
          type: boolean
          default: false
        upcomingEndingDates:
          $ref: '#/components/schemas/Settings_notifications_upcomingEndingDates'
      example:
        upcomingEndingDates:
          toggled: true
          fiveDaysBefore: false
          twoWeeksBefore: true
          oneMonthBefore: false
        completeProjectSheets: true
        employeeSheetReminder: true
        individualSheets: false
...
  securitySchemes:
    devnest_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://me.devnest.ro/oauth/auth
          scopes:
            read:timesheets: Read timesheets
            read:projects: Read projects
            write:projects: Update projects
            read:employees: Read employees
            write:employees: Update employees
            read:allocations: Read allocations
            write:allocations: Update allocations
            read:history: Read history
            write:history: Update history
            read:feedback: Read feedback
            write:feedback: Update feedback
            read:settings: Read settings
            write:settings: Write settings