# go/model_new_employee.go
/*
 * DevNest Portal API (OpenAPI 3.0)
 *
 * The DevNest Portal API endpoints definitions based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.0
 * Contact: contact@devnest.ro
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger
import (
	"time"
)

type NewEmployee struct {

	UserId float64 `json:"userId,omitempty"`

	FirstName string `json:"firstName"`

	LastName string `json:"lastName"`

	CNP string `json:"CNP"`

	JobTitle string `json:"jobTitle"`

	City string `json:"city"`

	PhoneNumber string `json:"phoneNumber"`

	WorkEmail string `json:"workEmail"`

	PersonalEmail string `json:"personalEmail"`

	EmployeeRateType string `json:"employeeRateType,omitempty"`

	EmployeeRateValue float64 `json:"employeeRateValue,omitempty"`

	ContractType string `json:"contractType,omitempty"`

	Currency string `json:"currency,omitempty"`

	VacationDays float64 `json:"vacationDays,omitempty"`

	StartingDate time.Time `json:"startingDate,omitempty"`

	EndingDate time.Time `json:"endingDate,omitempty"`

	HasMedicalPackage bool `json:"hasMedicalPackage,omitempty"`

	Status bool `json:"status,omitempty"`
}